requirements.txt
paho-mqtt==1.4.0
pyzmq==18.1.0
